\name{fS.detect}
\alias{fS.detect}

\title{
Used in the function detect
}
\description{
Used in the function detect for calculating the sample size for a square lattice design.
It is used by the \code{\link[stats]{optimize}} function.
}

\author{
Jon Barry: Jon.Barry@cefas.co.uk
}

\seealso{\code{\link{detect}}}


