\name{fT.detect}
\alias{fT.detect}

\title{
Used in function detect
}
\description{
Used in the function detect for calculating the sample size for a triangular lattice design.
It is used by the \code{\link[stats]{optimize}} function.
}

\author{
Jon Barry: Jon.Barry@cefas.co.uk
}

\seealso{\code{\link{detect}}}