\name{mom.min.nb}

\alias{mom.min.nb}

\title{Minimising function for VFC MOM estimator assuming Negative Binomial counts.}

\description{
Used by the \code{\link[stats]{optimize}} function to obtain the method of moments estimator within the function
\code{\link{GVFCMOM}}.
}

\author{
Jon Barry: Jon.Barry@cefas.co.uk
}

\seealso{ \code{\link[stats]{optimize}}, \code{\link{GVFCMOM}} }