\name{n.min}
\alias{n.min}
\title{Minimising function used in \code{\link{precision}}.}
\description{
Used by the \code{\link[stats]{optimize}} function to obtain the correct sample size within the function
\code{\link{precision}} given that the t-distributions is also a function of sample size.
}


\author{
Jon Barry: Jon.Barry@cefas.co.uk
}

\seealso{ \code{\link[stats]{optimize}}, \code{\link{precision}} }